+++ docs/pages/changelog.rst
@@ -27,6 +27,7 @@ Fixed
 - Added validation for closing Lakeshore |lakeshore335|.
 - Hidden internal text visible when selecting plugins in Plugin dialog |pluginDialog|.
 - Fixed displayed number of maximal errors for devices in the settings tooltip.
+- Resetting error count of virtual channels when reinitializing communication.
 
 Version 0.8.2 2025-06-23
 ========================